[WebApp][WIP] New categories page
