+++ apps/webapp/src/routes/_app/categories/index.tsx
@@ -1,183 +1,97 @@
-import { createFileRoute, useNavigate } from "@tanstack/react-router";
-
-import { useMediaQuery } from "@mantine/hooks";
+import { createFileRoute } from "@tanstack/react-router";
 import {
-  createCategory,
-  updateCategory as serverUpdateCategory,
-  deleteCategory as serverDeleteCategory,
-} from "@/features/categories/api/categories.api";
-import {
-  CategoriesDetailsModal,
-  CategoryFormData,
-} from "@/features/categories/CategoryDetailsModal";
-import { z } from "zod";
-import { Button, Modal, Stack, Text } from "@mantine/core";
-import { CategoriesList } from "@/features/categories/components/CategoriesList/CategoriesList";
-import { useCategories } from "@guallet/api-react";
-
-const categoriesSearchSchema = z.object({
-  modal: z.enum(["create", "edit"]).optional(),
-  categoryId: z.string().optional(),
-  parentId: z.string().optional(),
-});
+  Accordion,
+  ActionIcon,
+  Badge,
+  Button,
+  Divider,
+  Group,
+  Stack,
+  Text,
+  Tooltip,
+} from "@mantine/core";
+import { AppCategory, useGroupedCategories } from "@guallet/api-react";
+import { AppScreen } from "@/components/Layout/AppScreen";
+import { AppSection } from "@/components/Cards/AppSection";
+import { CategoryAvatar } from "@/components/Categories/CategoryAvatar";
+import { useState } from "react";
+import { IconChevronDown, IconChevronRight } from "@tabler/icons-react";
 
 export const Route = createFileRoute("/_app/categories/")({
   component: CategoriesPage,
-  validateSearch: categoriesSearchSchema,
 });
 
 function CategoriesPage() {
-  const { categories } = useCategories();
-
-  const { modal, categoryId, parentId } = Route.useSearch();
-
-  const isMobile = useMediaQuery("(max-width: 50em)");
-  const isCategoryDetailsModalOpen = modal !== undefined;
-
-  const navigate = useNavigate({ from: Route.fullPath });
-
-  async function saveCategory(data: CategoryFormData) {
-    console.log("Create new category", data);
-    if (data.parentId === null && parentId) {
-      data.parentId = parentId;
-    }
-    await createCategory(data);
-    hideModal(true);
-  }
-
-  async function updateCategory(id: string, data: CategoryFormData) {
-    console.log("Update from cat form", id);
-    try {
-      await serverUpdateCategory(id, data);
-    } catch (e) {
-      console.error("Error Deleting the remote category", e);
-    } finally {
-      hideModal(true);
-    }
-  }
-
-  async function deleteCategory(id: string) {
-    console.log("OnDelete from cat form", id);
-    try {
-      await serverDeleteCategory(id);
-    } catch (e) {
-      console.error("Error Deleting the remote category", e);
-    } finally {
-      hideModal(true);
-    }
-  }
-
-  function showModal(args: {
-    categoryId: string | null;
-    parentId: string | null;
-  }) {
-    const { categoryId, parentId } = args;
-
-    navigate({
-      search: () => ({
-        modal: categoryId ? "edit" : "create",
-        categoryId: categoryId ?? undefined,
-        parentId: parentId ?? undefined,
-      }),
-    });
-  }
-
-  function hideModal(forceRefresh: boolean = false) {
-    navigate({
-      search: () => ({
-        modal: undefined,
-        categoryId: undefined,
-        parentId: undefined,
-      }),
-    });
-
-    if (forceRefresh) {
-      Route.router?.invalidate();
-    }
-  }
-
-  if (categories.length == 0) {
-    return (
-      <Stack>
-        <Text>You don't have any category yet</Text>
-        <Button
-          onClick={() =>
-            showModal({
-              categoryId: null,
-              parentId: null,
-            })
-          }
-        >
-          Create your first category
-        </Button>
-      </Stack>
-    );
-  }
+  const { categories, isLoading } = useGroupedCategories();
 
   return (
-    <>
-      <Modal
-        centered
-        opened={isCategoryDetailsModalOpen}
-        overlayProps={{ opacity: 0.55, blur: 3 }}
-        onClose={() => hideModal()}
-        fullScreen={isMobile}
-        title={
-          !parentId
-            ? "Add new category"
-            : `Add subcategory for ${
-                categories.find((x) => x.id === parentId)?.name
-              }`
-        }
-      >
-        <CategoriesDetailsModal
-          category={categories.find((x) => x.id === categoryId) ?? null}
-          onCancel={hideModal}
-          onSave={(x) => {
-            if (parentId) {
-              x.parentId = parentId;
-            }
-            console.log("OnSubmit cat form", x);
-            saveCategory(x);
-          }}
-          onUpdate={(category, formData) => {
-            updateCategory(category.id, formData);
-          }}
-          onDelete={(x) => {
-            deleteCategory(x.id);
-          }}
-        />
-      </Modal>
+    <AppScreen isLoading={isLoading}>
       <Stack>
-        <Button
-          onClick={() =>
-            showModal({
-              categoryId: null,
-              parentId: null,
-            })
-          }
-        >
+        <Button onClick={() => console.log("Add new parent category")}>
           Add new parent category
         </Button>
-        <CategoriesList
-          categories={categories}
-          onAddSubcategory={(parent) => {
-            showModal({
-              categoryId: null,
-              parentId: parent.id,
-            });
-          }}
-          onEdit={(x) => {
-            showModal({
-              categoryId: x.id,
-              parentId: null,
-            });
-          }}
-          onDelete={(x) => {
-            deleteCategory(x.id);
-          }}
-        />
+        <Accordion>
+          {categories.map((category) => (
+            <CategorySection key={category.id} category={category} />
+          ))}
+        </Accordion>
       </Stack>
-    </>
+    </AppScreen>
+  );
+}
+
+function CategorySection({ category }: { category: AppCategory }) {
+  const [isExpanded, setExpanded] = useState(false);
+
+  return (
+    <AppSection title="">
+      <Group>
+        {isExpanded ? (
+          <ActionIcon
+            variant="transparent"
+            onClick={() => {
+              setExpanded(false);
+            }}
+          >
+            <IconChevronDown />
+          </ActionIcon>
+        ) : (
+          <ActionIcon
+            variant="transparent"
+            onClick={() => {
+              setExpanded(true);
+            }}
+          >
+            <IconChevronRight />
+          </ActionIcon>
+        )}
+
+        <CategoryAvatar categoryId={category.id} color={category.colour} />
+        <Text>{category.name}</Text>
+        <Tooltip
+          label={
+            category.subCategories.length === 0
+              ? "No subcategories"
+              : `${category.subCategories.length} sub-categories`
+          }
+        >
+          <Badge circle>{category.subCategories.length}</Badge>
+        </Tooltip>
+      </Group>
+      {isExpanded && (
+        <Stack>
+          {category.subCategories.length > 0 && (
+            <Divider label="Sub-categories" labelPosition="center" />
+          )}
+          {category.subCategories.map((subCategory) => (
+            <Group key={subCategory.id}>
+              <CategoryAvatar categoryId={subCategory.id} />
+              <Text>{subCategory.name}</Text>
+            </Group>
+          ))}
+          <Button variant="outline">Add sub-category</Button>
+        </Stack>
+      )}
+    </AppSection>
   );
 }
